Add hooks for other plugins
Add 3 filters:
 * authmgr_action_capability_map_filter
 * authmgr_anon_capabilities_filter
 * authmgr_role_capabilities_filter

authmgr_action_capability_map_filter:
This filter allows plugins to change the action/capability map used by
Authmgr to map Yourls Ajax actions to Authmgr capabilities. This is
useful for plugins that add additional Ajax actions.
The map is an associative array where the key is the action id, and the
value is the capability name.

authmgr_anon_capabilities_filter:
This filter allows plugins to change the capabilities of the anonymous
user. This is useful for plugins that want to add Authmgr capabilities.

authmgr_role_capabilities_filter:
This filter allows plugins to change the capabilities users an assigned
role. This is useful for plugins that want to add Authmgr capabilities.